requirements.txt
Flask
scikit-learn
pandas
numpy
matplotlib
torch
torchvision
timm
fastapi
uvicorn
requests
pillow
python-multipart